checkpoint; prior than direct matching boundary edges
svn path=/vcell/branches/MovingBoundary/; revision=14525

+++ Solver/include/VoronoiResult.h
@@ -7,7 +7,7 @@ namespace spatial {
 	/**
 	* implementation type for boost::polygon::voronoi routines
 	*/
-	typedef signed int VoronoiType;
+	typedef signed long VoronoiType;
 
 	typedef GhostPoint<VoronoiType,2> VoronoiGhostPoint;
